♻️ assertResponseEqual: dupped, moved to testutils
- Two test files definded their own version
and one of them was inferior.
- Moved to testutils
